fix: crash on library backup export and translate page caching
org.json.JSONObject.put(String, float) does not exist on Android's
platform JSONObject (only the desktop reference implementation has that
overload) - calling it throws NoSuchMethodError at runtime, which is an
Error, not an Exception, so it wasn't caught by any of the surrounding
try/catch(Exception) blocks and crashed the app outright.

This hit two real paths: TranslateRepository.serialize() (crashes the
first time a translated page gets cached) and BackupManager's library
export, which put chapterNumber - present on every chapter - the same
way, meaning "Zálohovat knihovnu" crashed for any library with at least
one chapter. Fixed by converting all four call sites to Double, which
Android's JSONObject does support.

+++ app/src/main/kotlin/com/haise/jiyu/translate/TranslateRepository.kt
@@ -146,10 +146,12 @@ class TranslateRepository @Inject constructor(
             arr.put(JSONObject().apply {
                 put("orig", b.originalText)
                 put("trans", b.translatedText)
-                put("l", b.leftF)
-                put("t", b.topF)
-                put("r", b.rightF)
-                put("b", b.bottomF)
+                // put(String, float) na Android org.json.JSONObject neexistuje (jen desktopová
+                // verze knihovny) -> NoSuchMethodError za běhu. Double overload existuje vždy.
+                put("l", b.leftF.toDouble())
+                put("t", b.topF.toDouble())
+                put("r", b.rightF.toDouble())
+                put("b", b.bottomF.toDouble())
             })
         }
     }.toString()
